+++ src/register_address.rs
@@ -1,6 +1,6 @@
 use crate::types::{
-    AccelOutputDataRate, AccelScale, AccelerometerId, MagOutputDataRate, MagnetometerId,
-    StatusFlags,
+    AccelOutputDataRate, AccelScale, AccelerometerId, FifoMode, Interrupt, MagOutputDataRate,
+    MagnetometerId, StatusFlags,
 };
 
 pub trait RegRead<D = u8> {
@@ -157,6 +157,7 @@ register! {
 
 register! {
   /// CTRL_REG3_A
+  #[derive(Default)]
   pub struct CtrlReg3A: 0x22 {
     const I1_CLICK   = 0b10000000;
     const I1_AOI1    = 0b01000000;
@@ -166,6 +167,32 @@ register! {
     const I1_WTM     = 0b00000100;
     const I1_OVERRUN = 0b00000010;
   }
+}
+
+impl CtrlReg3A {
+    pub const fn with_interrupt(self, interrupt: Interrupt) -> Self {
+        match interrupt {
+            Interrupt::Click => self.union(Self::I1_CLICK),
+            Interrupt::Aoi1 => self.union(Self::I1_AOI1),
+            Interrupt::Aoi2 => self.union(Self::I1_AOI2),
+            Interrupt::DataReady1 => self.union(Self::I1_DRDY1),
+            Interrupt::DataReady2 => self.union(Self::I1_DRDY2),
+            Interrupt::FifoWatermark => self.union(Self::I1_WTM),
+            Interrupt::FifoOverrun => self.union(Self::I1_OVERRUN),
+        }
+    }
+
+    pub const fn without_interrupt(self, interrupt: Interrupt) -> Self {
+        match interrupt {
+            Interrupt::Click => self.difference(Self::I1_CLICK),
+            Interrupt::Aoi1 => self.difference(Self::I1_AOI1),
+            Interrupt::Aoi2 => self.difference(Self::I1_AOI2),
+            Interrupt::DataReady1 => self.difference(Self::I1_DRDY1),
+            Interrupt::DataReady2 => self.difference(Self::I1_DRDY2),
+            Interrupt::FifoWatermark => self.difference(Self::I1_WTM),
+            Interrupt::FifoOverrun => self.difference(Self::I1_OVERRUN),
+        }
+    }
 }
 
 register! {
@@ -208,6 +235,7 @@ impl CtrlReg4A {
 
 register! {
   /// CTRL_REG5_A
+  #[derive(Default)]
   pub struct CtrlReg5A: 0x24 {
     const BOOT     = 0b10000000;
     const FIFO_EN  = 0b01000000;
@@ -237,6 +265,7 @@ register! {
 
 register! {
   /// FIFO_CTRL_REG_A
+  #[derive(Default)]
   pub struct FifoCtrlRegA: 0x2E {
     const FM1  = 0b10000000;
     const FM0  = 0b01000000;
@@ -246,7 +275,31 @@ register! {
     const FTH2 = 0b00000100;
     const FTH1 = 0b00000010;
     const FTH0 = 0b00000001;
+
+    const FM = Self::FM1.bits | Self::FM0.bits;
+    const FTH = Self::FTH4.bits | Self::FTH3.bits | Self::FTH2.bits | Self::FTH1.bits | Self::FTH0.bits;
   }
+}
+
+impl FifoCtrlRegA {
+    pub const fn with_mode(self, mode: FifoMode) -> Self {
+        match mode {
+            FifoMode::Bypass => self.difference(Self::FM),
+            FifoMode::Fifo => self.difference(Self::FM1).union(Self::FM0),
+            FifoMode::Stream => self.union(Self::FM1).difference(Self::FM0),
+            FifoMode::StreamToFifo => self.union(Self::FM),
+        }
+    }
+
+    pub const fn with_full_threshold(self, n: u8) -> Self {
+        let n = if n > Self::FTH.bits {
+            Self::FTH.bits
+        } else {
+            n
+        };
+        self.difference(Self::FTH)
+            .union(Self::from_bits_truncate(n))
+    }
 }
 
 register! {
@@ -416,6 +469,30 @@ mod tests {
         check_odr(AccelOutputDataRate::Hz100, 0b0101);
         check_odr(AccelOutputDataRate::Hz200, 0b0110);
         check_odr(AccelOutputDataRate::Hz400, 0b0111);
+    }
+
+    #[test]
+    fn ctrl_reg_3_a() {
+        let ctrl = CtrlReg3A::default();
+        let ctrl_all = CtrlReg3A::from_bits_truncate(0b11111110);
+
+        let mut bits = 0b10000000;
+        for interrupt in [
+            Interrupt::Click,
+            Interrupt::Aoi1,
+            Interrupt::Aoi2,
+            Interrupt::DataReady1,
+            Interrupt::DataReady2,
+            Interrupt::FifoWatermark,
+            Interrupt::FifoOverrun,
+        ] {
+            assert_eq!(ctrl.with_interrupt(interrupt).bits(), bits,);
+            assert_eq!(
+                ctrl_all.without_interrupt(interrupt).bits(),
+                (!bits) & 0b11111110
+            );
+            bits >>= 1;
+        }
     }
 
     #[test]