+++ src/types.rs
@@ -588,3 +588,35 @@ impl Temperature {
         (self.unscaled() as f32) / 256.0 + Self::DEFAULT
     }
 }
+
+/// A FIFO mode.
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum FifoMode {
+    /// Bypass mode
+    Bypass,
+    /// FIFO mode
+    Fifo,
+    /// Stream mode
+    Stream,
+    /// Stream-to-FIFO mode
+    StreamToFifo,
+}
+
+/// An interrupt.
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum Interrupt {
+    /// AOI1 interrupt on INT1 pin.
+    Aoi1,
+    /// AOI2 interrupt on INT1 pin.
+    Aoi2,
+    /// CLICK interrupt on INT1 pin.
+    Click,
+    /// DRDY1 interrupt on INT1 pin.
+    DataReady1,
+    /// DRDY2 interrupt on INT1 pin.
+    DataReady2,
+    /// FIFO overrun interrupt on INT1 pin.
+    FifoOverrun,
+    /// FIFO watermark interrupt on INT1 pin.
+    FifoWatermark,
+}